compare the minimum packet size on packet size determination
when we compare the packet size configured either with 1) ccPacketSize
(ClientConfig) or 2) accPacketSize (received pkt size at a server), it
compared with defaultQUICPacketSize, which is 1350 or 1330 (ipv6 or
ipv4).  As a result, we cannot configure ccPacketSize less than that
value.

This commit fixes to use defaultQUICPacketSize, 1200, which is the
smallest allowed maximum datagram size defined in RFC9000, to allow to
configure ccPacketSize less than 1350/1300 bytes.

Signed-off-by: Hajime Tazaki <thehajime@gmail.com>